+++ enclone_main/src/lib.rs
@@ -2,6 +2,7 @@
 
 pub mod determine_ref;
 pub mod doublets;
+pub mod filter_umi;
 pub mod inconsistent;
 pub mod main_enclone;
 pub mod merge_onesies;
